+++ include/nighthawk/client/factories.h
@@ -32,22 +32,6 @@ public:
                                     RequestSource& request_generator) const PURE;
 };
 
-class SequencerFactory {
-public:
-  virtual ~SequencerFactory() = default;
-  virtual SequencerPtr create(Envoy::TimeSource& time_source, Envoy::Event::Dispatcher& dispatcher,
-                              BenchmarkClient& benchmark_client,
-                              TerminationPredicatePtr&& termination_predicate,
-                              Envoy::Stats::Scope& scope,
-                              const Envoy::MonotonicTime scheduled_starting_time) const PURE;
-};
-
-class StatisticFactory {
-public:
-  virtual ~StatisticFactory() = default;
-  virtual StatisticPtr create() const PURE;
-};
-
 class OutputFormatterFactory {
 public:
   virtual ~OutputFormatterFactory() = default;
@@ -56,21 +40,4 @@ public:
 };
 
 } // namespace Client
-
-class RequestSourceFactory {
-public:
-  virtual ~RequestSourceFactory() = default;
-  virtual RequestSourcePtr create(const Envoy::Upstream::ClusterManagerPtr& cluster_manager,
-                                  Envoy::Event::Dispatcher& dispatcher, Envoy::Stats::Scope& scope,
-                                  absl::string_view service_cluster_name) const PURE;
-};
-
-class TerminationPredicateFactory {
-public:
-  virtual ~TerminationPredicateFactory() = default;
-  virtual TerminationPredicatePtr
-  create(Envoy::TimeSource& time_source, Envoy::Stats::Scope& scope,
-         const Envoy::MonotonicTime scheduled_starting_time) const PURE;
-};
-
 } // namespace Nighthawk
