+++ include/nighthawk/common/sequencer.h
@@ -6,9 +6,12 @@
 
 #include "envoy/common/pure.h"
 
+#include "nighthawk/common/operation_callback.h"
 #include "nighthawk/common/statistic.h"
 
 namespace Nighthawk {
+
+using SequencerTarget = std::function<bool(OperationCallback)>;
 
 /**
  * Abstract Sequencer interface.
