+++ source/client/client_worker_impl.cc
@@ -34,13 +34,17 @@ ClientWorkerImpl::ClientWorkerImpl(Envoy::Api::Api& api, Envoy::ThreadLocal::Ins
       benchmark_client_(benchmark_client_factory.create(
           api, *dispatcher_, *worker_number_scope_, cluster_manager, http_tracer_,
           fmt::format("{}", worker_number), *request_generator_)),
-      phase_(std::make_unique<PhaseImpl>(
-          "main",
-          sequencer_factory_.create(*time_source_, *dispatcher_, *benchmark_client_,
-                                    termination_predicate_factory_.create(
-                                        *time_source_, *worker_number_scope_, starting_time),
-                                    *worker_number_scope_, starting_time),
-          true)),
+      phase_(
+          std::make_unique<PhaseImpl>("main",
+                                      sequencer_factory_.create(
+                                          *time_source_, *dispatcher_,
+                                          [this](CompletionCallback f) -> bool {
+                                            return benchmark_client_->tryStartRequest(std::move(f));
+                                          },
+                                          termination_predicate_factory_.create(
+                                              *time_source_, *worker_number_scope_, starting_time),
+                                          *worker_number_scope_, starting_time),
+                                      true)),
       hardcoded_warmup_style_(hardcoded_warmup_style) {}
 
 void ClientWorkerImpl::simpleWarmup() {
