+++ source/common/sequencer_impl.h
@@ -4,7 +4,6 @@
 #include "envoy/common/time.h"
 #include "envoy/event/dispatcher.h"
 
-#include "nighthawk/common/operation_callback.h"
 #include "nighthawk/common/platform_util.h"
 #include "nighthawk/common/rate_limiter.h"
 #include "nighthawk/common/sequencer.h"
@@ -23,8 +22,6 @@ using namespace std::chrono_literals;
 constexpr std::chrono::microseconds NighthawkTimerResolution = 25us;
 
 } // namespace
-
-using SequencerTarget = std::function<bool(OperationCallback)>;
 
 using namespace Envoy; // We need this because of macro expectations.
