+++ test/client_worker_test.cc
@@ -14,12 +14,14 @@
 
 #include "client/client_worker_impl.h"
 
-#include "test/mocks.h"
 #include "test/mocks/client/mock_benchmark_client.h"
 #include "test/mocks/client/mock_benchmark_client_factory.h"
 #include "test/mocks/common/mock_request_source.h"
+#include "test/mocks/common/mock_request_source_factory.h"
 #include "test/mocks/common/mock_sequencer.h"
+#include "test/mocks/common/mock_sequencer_factory.h"
 #include "test/mocks/common/mock_termination_predicate.h"
+#include "test/mocks/common/mock_termination_predicate_factory.h"
 
 #include "gtest/gtest.h"
