--- test/mocks.cc
@@ -1,15 +1,0 @@
-#include "test/mocks.h"
-
-#include "gmock/gmock.h"
-
-namespace Nighthawk {
-
-MockSequencerFactory::MockSequencerFactory() = default;
-
-MockStatisticFactory::MockStatisticFactory() = default;
-
-MockRequestSourceFactory::MockRequestSourceFactory() = default;
-
-MockTerminationPredicateFactory::MockTerminationPredicateFactory() = default;
-
-} // namespace Nighthawk
