+++ test/mocks/common/mock_request_source_factory.cc
@@ -1,0 +1,7 @@
+#include "test/mocks/common/mock_request_source_factory.h"
+
+namespace Nighthawk {
+
+MockRequestSourceFactory::MockRequestSourceFactory() = default;
+
+} // namespace Nighthawk
