+++ test/mocks/common/mock_sequencer_factory.cc
@@ -1,0 +1,7 @@
+#include "test/mocks/common/mock_sequencer_factory.h"
+
+namespace Nighthawk {
+
+MockSequencerFactory::MockSequencerFactory() = default;
+
+} // namespace Nighthawk
