+++ test/mocks/common/mock_termination_predicate_factory.cc
@@ -1,0 +1,7 @@
+#include "test/mocks/common/mock_termination_predicate_factory.h"
+
+namespace Nighthawk {
+
+MockTerminationPredicateFactory::MockTerminationPredicateFactory() = default;
+
+} // namespace Nighthawk
