+++ test/process_test.cc
@@ -13,7 +13,6 @@
 #include "client/process_impl.h"
 
 #include "test/client/utility.h"
-#include "test/mocks.h"
 
 #include "gtest/gtest.h"
